+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased] - yyyy-mm-dd
 
+## [3.8.0] - 2024-01-31
+
+### Added
+
+- New dashboard panel showing interesting KPIs for a users's character, e.g. total wallet ISK or total skillpoints (#154)
+
+### Changed
+
+- Add optional storing of raw ESI data for contract items and bids
+- Add tests to prove that contract items are only fetched once (see also #60)
+
 ## [3.7.5] - 2024-01-25
 
 >**IMPORTANT**: When updating from a version prior to 3.7.2, 3.6.0, 3.4.0 or 3.0.0 please see the update notes for those versions first!